+++ Task_List_Kata/java/src/main/java/com/codurance/training/tasks/service/ShowByDateService.java
@@ -1,0 +1,10 @@
+package com.codurance.training.tasks.service;
+
+import com.codurance.training.tasks.Task;
+
+import java.util.List;
+import java.util.Map;
+
+public interface ShowByDateService {
+    public void showByDate(Map<String, List<Task>> tasks);
+}
