+++ Task_List_Kata/java/src/main/java/com/codurance/training/tasks/service/ShowByProjectService.java
@@ -1,0 +1,11 @@
+package com.codurance.training.tasks.service;
+
+import com.codurance.training.tasks.Task;
+
+import java.util.List;
+import java.util.Map;
+
+public interface ShowByProjectService {
+
+    public void showByProject(Map<String, List<Task>> tasks);
+}
