+++ Task_List_Kata/java/src/main/java/com/codurance/training/tasks/service/impl/ShowDueTaskServiceImpl.java
@@ -1,0 +1,40 @@
+package com.codurance.training.tasks.service.impl;
+
+import com.codurance.training.tasks.Task;
+import com.codurance.training.tasks.service.ShowDueTaskService;
+
+import java.io.PrintWriter;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.List;
+import java.util.Map;
+
+public class ShowDueTaskServiceImpl implements ShowDueTaskService {
+
+    private final PrintWriter out;
+
+    public ShowDueTaskServiceImpl(PrintWriter writer) {
+        this.out = writer;
+    }
+
+    @Override
+    public void showDueTodayTasks(Map<String, List<Task>> tasks) {
+
+        Date today = new Date();
+
+        for (Map.Entry<String, List<Task>> project : tasks.entrySet()) {
+            out.println(project.getKey());
+            for (Task task : project.getValue()) {
+                if(task.getDeadline() != null && parseDate(task.getDeadline()).equals(parseDate(today)))
+                    out.printf("    [%c] %s: %s %s%n", (task.isDone() ? 'x' : ' '), task.getId(), task.getDescription(), parseDate(task.getDeadline()));
+            }
+            out.println();
+        }
+    }
+
+    private String parseDate(Date date) {
+        SimpleDateFormat formatter = new SimpleDateFormat("dd-MM-yyyy");
+        return formatter.format(date);
+    }
+
+}
